+++ host/usb/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this component will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Device initiated remote wakeup (https://github.com/espressif/esp-usb/pull/298)
+
 ## [1.1.0] - 2025-12-04
 
 ### Added